Move pdb saving to DebugingWidget
